requirements.txt
langchain
langchain-google-genai
langchain-community
langgraph
python-dotenv
pydantic 
tavily-python
chromadb
pypdf